change to 20 tiles displayed per page
