+++ src/features/PersonList/popularPeopleSaga.js
@@ -8,21 +8,7 @@ import {
 
 function* fetchPopularPeopleHandler({ payload: pageNumber }) {
   try {
-    const numberOfPagesToGetFromApi = 9;
-    const highestNumberOfPageToGetFromApi =
-      pageNumber + numberOfPagesToGetFromApi - 1;
-    let pageNumberForApi = pageNumber;
-    let popularPeopleData = [];
-
-    for (
-      pageNumberForApi;
-      pageNumberForApi <= highestNumberOfPageToGetFromApi;
-      pageNumberForApi++
-    ) {
-      let dataFromApi = yield call(getPopularPeople, pageNumberForApi);
-      popularPeopleData = popularPeopleData.concat(dataFromApi);
-    }
-
+    const popularPeopleData = yield call(getPopularPeople, pageNumber);
     yield put(fetchPopularPeopleSuccess(popularPeopleData));
   } catch (error) {
     yield put(fetchPopularPeopleError());
